undo networking and return to one standalone binary
The goals actually never demanded networking, but I added it,
thinking it would be fun.
However, it made me reconsider using Go for at least the client and server,
which is ridiculous, given that I recently moved past it.
So instead of going back to Go, I decided to drop networking for now.
Also, having to implement any user action in both client and server feels so
indirect and slow.
Plus, testing is kind of a nightmare,
because whenever I do something wrong with the socket,
my OS may just cut me off.
All of these cause a net loss of fun, and we can't have that.