replace gtk client implementation with tk
This also renames the package to reflect the backend rather than the target.
Also remove cross_platform to clean up.
Removing cross_platform is done because it's obsolete.
With core and extra being in C now, it would need more CGO,
taking performance hits. That, and I don't want to bind that.