revive ebiten client and port libs to Go
There are also some architectural differences.
mat is now a submodule of core,
to bring back the proper namespace use,
which is nicer to read.
The ebiten client also got renamed,
and changed to fit the new libs.
Also fix an old Makefile mistake from prior renaming attempt,
that messed with ldflags.